Live deployment - reduce job downtime by spawning a new AM for new deployment while job is already running from previous deployment. (#3)
Live deployment - reduce job downtime by spawning a new AM for new deployment while job is already running from previous deployment.

This changeset introduces logic for the live deployment feature in which we spawn a new AM while the existing deployment of the job is still running. However we want to make sure that at any given time there are only one set of containers running for the job. The new AM comes up and when it is ready to spawn up the containers, it kills existing deployment of the job (which would kill existing AM and existing containers for the job), waits for it to be moved to KILLED state and then continues with requesting containers from RM.

This feature is behind a flag which is based on config of the job.

Other changes in this changeset include -
1: change in clientHelper code so that it returns active app ids for a job in sorted order based on start time. Earlier this method would have returned only one job always, since there is only on job in running state. However now it can return two when the job is being deployed.
2: Small refactor in SamzaYarnAppMasterService code to move “write of AM info on coordinator stream” to a new method and use that method if the live deployment feature flag is not set, which means there is no change in flow if "live deployment" flag is not set. If live deployment flag is set, then we write AM info on coordinator stream after the existing AM and all containers are killed.

---------

Co-authored-by: Rahul Yadav <rahyadav@rahyadav-ld2.linkedin.biz>