Changing retention policy of kafka topics for a SPCP job from 28 days to 14 days.
GDPR and DMA require us to delete member data within 30 days of member account closure.
However, there is a 14 day lag between when a user hits delete and LinkedIn closes
their account. Which means we need to have retention of 14 days of member data.

This change only applies to scenarios where the retention was 28 days.